add rep_tests cases & fix meta's server_state::update_configuration()

+++ build.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+CMAKE_OPTIONS="$CMAKE_OPTIONS -DCMAKE_BUILD_TYPE=Debug"
+CMAKE_OPTIONS="$CMAKE_OPTIONS -DCMAKE_C_COMPILER=gcc -DCMAKE_CXX_COMPILER=g++"
+
 # You can specify customized boost by defining BOOST_DIR.
 # Install boost like this:
 #   wget http://downloads.sourceforge.net/project/boost/boost/1.54.0/boost_1_54_0.zip?r=&ts=1442891144&use_mirror=jaist
@@ -13,23 +16,38 @@
 if [ -n "$BOOST_DIR" ]
 then
     echo "Use customized boost: $BOOST_DIR"
-    BOOST_OPTIONS="-DBoost_NO_BOOST_CMAKE=ON -DBOOST_ROOT=$BOOST_DIR -DBoost_NO_SYSTEM_PATHS=ON"
-else
-    BOOST_OPTIONS=""
+    CMAKE_OPTIONS="$CMAKE_OPTIONS -DBoost_NO_BOOST_CMAKE=ON -DBOOST_ROOT=$BOOST_DIR -DBoost_NO_SYSTEM_PATHS=ON"
 fi
 
-if [ $# -eq 1 -a "$1" == "true" ]
+if [ -d "builder" -a $# -eq 1 -a "$1" == "true" ]
 then
     echo "Clear builder..."
     rm -rf builder
+fi
+
+if [ ! -d "builder" ]
+then
+    echo "Running cmake..."
     mkdir -p builder
     cd builder
-    cmake .. -DCMAKE_INSTALL_PREFIX=`pwd`/output -DCMAKE_BUILD_TYPE=Debug $BOOST_OPTIONS
+    cmake .. -DCMAKE_INSTALL_PREFIX=`pwd`/output $CMAKE_OPTIONS
+    if [ $? -ne 0 ]
+    then
+        echo "ERROR: cmake failed"
+        exit -1
+    fi
     cd ..
 fi
 
 cd builder
-make -j4
-make install
+echo "Building..."
+make install -j8
+if [ $? -ne 0 ]
+then
+    echo "ERROR: build failed"
+    exit -1
+else
+    echo "Build succeed"
+fi
 cd ..
